issue 22: adafruit backpack demos package update
