+++ robo4j-hw-rpi/src/examples/java/com/robo4j/hw/rpi/i2c/adafruitbackpack/BiColor8x8MatrixExample.java
@@ -15,7 +15,12 @@
  * along with Robo4J. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.robo4j.hw.rpi.i2c.adafruitoled;
+package com.robo4j.hw.rpi.i2c.adafruitbackpack;
+
+import com.robo4j.hw.rpi.i2c.adafruitoled.BiColor;
+import com.robo4j.hw.rpi.i2c.adafruitoled.BiColor8x8MatrixDevice;
+import com.robo4j.hw.rpi.i2c.adafruitoled.MatrixRotation;
+import com.robo4j.hw.rpi.i2c.adafruitoled.PackElement;
 
 import java.util.concurrent.TimeUnit;
